% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simjm.R
\name{plot.simjm}
\alias{plot.simjm}
\title{Plot method for simjm simulated data frames}
\usage{
\method{plot}{simjm}(x, type = c("all", "trajectory", "basehaz", "hazard"),
  m = NULL)
}
\arguments{
\item{x}{An object of class \code{"simjm"}, which is a list of
data frames returned by a call to \code{\link{simjm}}.}

\item{type}{The type of plot to produce. Type \code{"trajectory"}
produces a plot of the mean trajectory for each longitudinal
outcome, or if \code{m} is specified then the mean trajectory
for the \emph{m}th longitudinal outcome. Type \code{"basehaz"}
produces a plot of the hazard function evaluated at the mean
value of the baseline covariates and a zero contribution from
the association structure. Type \code{"hazard"} produces a
plot of the hazard function evaluated at the mean value of the
baseline covariates and at the current mean value (i.e. the
value at time \emph{t}) for each of the longitudinal outcome(s).
The default \code{"all"} produces a plot matrix with all of the
above plot types.}

\item{m}{A positive integer. If \code{NULL} then plots are shown
for all the longitudinal outcomes (only relevant if the data
was simulated for a multivariate joint model). Otherwise, one
can specify a specific longitudinal outcome to produce the plots
for. Note that is ignored if \code{type = "basehaz"} or
\code{type = "hazard"}.}
}
\value{
A ggplot object, which can be further customised using
  the \strong{ggplot2} package.
}
\description{
Plot the underlying longitudinal trajectory and hazard function
for the data generating model underpinning a \code{\link{simjm}}
call. The plots are deterministic functions that use the true
parameter values from the data generating model. These true
parameter values are stored as an attribute in the object returned
by a call to \code{\link{simjm}} (which is a list of data frames
with class \code{c("simjm", "list")}).
}
