% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traj.R
\name{plot_traj}
\alias{plot_traj}
\title{Plot observed longitudinal trajectories using facets}
\usage{
plot_traj(yvar = "Yij_1", data, data_filter = NULL, timevar = "tij",
  idvar = "id")
}
\arguments{
\item{yvar}{Name of the variable corresponding to the longitudinal biomarker.}

\item{data}{The data frame with simulated data for the longitudinal
submodel. The variable indicating the individual should be called "id",
as is the case in the data frames returned by \code{\link{simjm}}.}

\item{data_filter}{Optionally, a condition wrapped in quotes indicating the
subset of individuals to include in the plot (for example we could use
\code{data_filter = "id <= 20"} to include plots the first 20 individuals).
If \code{NULL} then a separate facet is included for every individual in
\code{data}.}

\item{timevar}{Name of the variable corresponding to time.}

\item{idvar}{Name of the variable identifying individuals.}
}
\value{
A \code{ggplot} object that can be further customised using the
  \pkg{ggplot2} package.
}
\description{
This is a convenience function that allows the user to easily plot the
observed longitudinal biomarker trajectory for each individual. Each individual
is shown in a separate facet in the plot. This function can be useful
to assess whether the "true" parameter values passed to \code{\link{simjm}}
are generating realistic longitudinal trajectories for the simulated
biomarkers.
}
\examples{
  dat <- simjm(n = 20, random_trajectory = "poly", b_sd = c(2,2,2), seed = 123)
  plot_traj("Yij_1", data = dat$Long1)
  plot_traj("Yij_1", data = dat$Long1, data_filter = "id <= 10")

}
