% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simjm.R
\name{simjm}
\alias{simjm}
\title{Simulate data for a univariate or multivariate joint model}
\usage{
simjm(n = 200, M = 1, random_trajectory = "linear", assoc = "etavalue",
  basehaz = c("weibull"), betaLong_intercept = 0, betaLong_binary = 1,
  betaLong_continuous = 1, betaLong_slope = 1, betaLong_poly1 = 1,
  betaLong_poly2 = 0.5, betaLong_aux = 1, betaEvent_intercept = -4,
  betaEvent_binary = 1, betaEvent_continuous = 0, betaEvent_assoc = 0.2,
  betaEvent_aux = 1.2, b_sd = c(2, 1), b_rho = -0.2, prob_Z1 = 0.5,
  mean_Z2 = 0, sd_Z2 = 1, max_yobs = 10, max_fuptime = 5,
  family = gaussian, clust_control = list(),
  seed = sample.int(.Machine$integer.max, 1), interval = c(1e-08, 200))
}
\arguments{
\item{n}{Number of individuals}

\item{M}{Number of longitudinal markers}

\item{random_trajectory}{The desired type of trajectory in the random effects
part of the longitudinal model. Can be \code{"linear"} (the default),
\code{"poly"}, or \code{"none"}. Can be a single character string, or a
character vector of length M (if a different trajectory type is to be used
for each longitudinal submodel). If \code{random_trajectory = "linear"} or
\code{random_trajectory = "none"} then a fixed effect linear
slope is also included in the longitudinal submodel. If
\code{random_trajectory = "poly"} then fixed effect linear and
quadratic terms are also included in the longitudinal submodel.}

\item{assoc}{A character string, or a character vector of length M,
specifying the desired type of association structure for
linking each longitudinal outcome to the hazard of the event.
Only one association structure type can be specified for each longitudinal
outcome. The options are: \code{"etavalue"}, \code{"etaslope"},
\code{"shared_b(1)"}, \code{"shared_coef(1)"}, \code{"shared_b(2)"},
\code{"shared_coef(2)"}.}

\item{basehaz}{The desired baseline hazard for the event submodel. Can be
\code{"weibull"}.}

\item{betaLong_intercept}{True intercept in the longitudinal submodel. Can
be a scalar or a vector of length M.}

\item{betaLong_binary}{True coefficient for the binary covariate in the
longitudinal submodel. Can be a scalar or a vector of length M.}

\item{betaLong_continuous}{True coefficient for the continuous covariate in the
longitudinal submodel. Can be a scalar or a vector of length M.}

\item{betaLong_slope}{True coefficient for the fixed effect slope in the
longitudinal submodel when \code{random_trajectory = "linear"} or
\code{random_trajectory = "none"}. Can be a scalar or a vector of length M.}

\item{betaLong_poly1}{True coefficient for the fixed effect linear term
in the longitudinal submodel when \code{random_trajectory = "poly"}. Can be
a scalar or a vector of length M.}

\item{betaLong_poly2}{True coefficient for the fixed effect quadratic term
in the longitudinal submodel when \code{random_trajectory = "poly"}. Can
be a scalar or a vector of length M.}

\item{betaLong_aux}{True parameter value for the auxiliary parameter in the
longitudinal submodel (sigma for Gaussian models, number of trials for
binomial models, size for negative binomial models, shape for Gamma models,
lambda for inverse Gaussian models). Can be a scalar or a vector of length M.}

\item{betaEvent_intercept}{True intercept term (log hazard scale) in the
event submodel. Only required for Weibull models.}

\item{betaEvent_binary}{True coefficient (log hazard ratio) for the binary
covariate in the event submodel.}

\item{betaEvent_continuous}{True coefficient (log hazard ratio) for the
continuous covariate in the event submodel.}

\item{betaEvent_assoc}{True association parameter (log hazard ratio) in the
event submodel. Can be a scalar or a vector of length M.}

\item{betaEvent_aux}{True parameter value(s) for the auxiliary parameter(s) in
the event submodel (shape parameter for Weibull models). Can be a scalar
or a vector, depending on how many auxiliary parameters are required for
specifying the baseline hazard.}

\item{b_sd}{A list, with each element of the list containing the vector of
standard deviations for the random effects relating to one grouping factor.}

\item{b_rho}{Correlation between the random effects. This is assumed to be
the correlation between all random effects within a grouping factor, and
the same correlation is used for each grouping factor.}

\item{prob_Z1}{Probability for the binary covariate included in each of the
submodels.}

\item{mean_Z2}{Mean of the (normally distributed) continuous covariate included
in each of the submodels.}

\item{sd_Z2}{Standard deviation of the (normally distributed) continuous
covariate included in each of the submodels.}

\item{max_yobs}{The maximum allowed number of longitudinal measurements. The
actual number of observed measurements will depend on the individuals event time.}

\item{max_fuptime}{The maximum follow up time in whatever the desired time
units are. This time will also be used as the censoring time (i.e. for
individuals who have a simulated survival time that is after \code{max_fuptime}).}

\item{family}{A family for the the longitudinal submodel, or for a multivariate
joint model this can be a list of families. See \code{\link[stats]{glm}} and
\code{\link[stats]{family}}.}

\item{clust_control}{A named list providing the arguments that are
necessary for simulating data where the first longitudinal submodel has
lower level clustering within individuals. The named list should contain
the following elements:
\describe{
\item{L}{Integer specifying the maximum number of lower level units
clustered within an individual. The actual number of lower level units
clustered within each individual is taken to be a random uniform (truncated
to an integer) on the range \code{[1,L+1]}.}
\item{assoc}{Character string specifying the method for combining information
across the lower level units clustered within an individual when forming the
association structure. Can be \code{"sum"} for specifying which indicates
the association structure should be based on a summation across the lower
level units clustered within an individual. Can be \code{"mean"} which
indicates that the association structure should be based on the mean
(i.e. average) taken across the lower level units clustered within an
individual.}
\item{u_sd}{Numeric vector providing the standard deviations of the random
effect at the cluster level.}
\item{random_trajectory}{The desired type of trajectory in the random effects
part of the longitudinal model at the cluster level. Can be \code{"none"} to
only include a random intercept at the cluster level, or otherwise
\code{"linear"} or \code{"poly"}.}
}}

\item{seed}{An optional \code{\link[=set.seed]{seed}}.}

\item{interval}{The interval over which to search for the
\code{\link[stats]{uniroot}} corresponding to each simulated event time.}
}
\value{
A list of data frames, one for each longitudinal biomarker (in long
  format) and one for the event time data. The returned object also has a
  number of attributes that record the values specified for the arguments in
  the \code{simjm} call (for example the "true" parameter values, the number
  of individuals, the number of longitudinal markers, and so on).
}
\description{
Returns a list of data frames containing data simulated from a joint model for
longitudinal and time-to-event data.
}
\details{
The \code{simjm} function returns data simulated under a joint
  longitudinal and time-to-event model. The joint model can be univariate
  (i.e. one longitudinal outcome) or multivariate (i.e. more than one
  longitudinal outcome). If more than one longitudinal outcome is specified
  then the longitudinal outcomes are assumed to be correlated via a joint
  multivariate normal distribution for the individual-level random effects.
  Each longitudinal outcome may have a different family and link function.
  The time-to-event model is assumed to be a parametric proportional hazards
  model with the baseline hazard specified via the \code{basehaz} argument.
  The (log) hazard of the event is assumed to be related to each longitudinal
  outcome via the association structure specified in the \code{assoc} argument.
  A different association may be specified for linking each longitudinal
  outcome to the hazard of the event.

  Note that by default \code{simjm} will simulate data for one longitudinal
  outcome (i.e. a univariate joint model). If we want to simulate data for
  a multivariate joint model, or we wish to use one of the non-default
  association structures (for example "etaslope"), then we may need to adjust
  the "true" parameters accordingly so that we still get realistic survival
  times (for example, by changing the "true" association parameter value).
}
\examples{
# Note that throughout the examples we specify 'n = 30' to
# ensure a small sample size so that the examples run quickly

# Simulate one longitudinal marker (we can just use the defaults):
simdat1 <- simjm(n = 30)

# Simulate two longitudinal markers and then check the
# true parameter values (stored as an attribute):
simdat2 <- simjm(M = 2, n = 30, betaEvent_assoc = 0.1)
attr(simdat2, "params")

# Simulate one longitudinal marker, using "etaslope"
# association structure:
simdat3 <- simjm(n = 30, assoc = "etaslope", betaEvent_assoc = 0.8)

# For one longitudinal marker, with a bernoulli outcome.
# (Note that 'betaLong_aux' specifies the number of trials
# for the binomial outcome).
simdat4 <- simjm(M = 1, n = 30,
                 betaLong_intercept = 1.3,
                 betaLong_binary = -0.6,
                 betaLong_continuous = -0.03,
                 betaLong_slope = 0.05,
                 b_sd = c(1, 0.05),
                 betaEvent_intercept = -9,
                 betaEvent_assoc = 0.3,
                 family = binomial(),
                 betaLong_aux = 1)

# For one longitudinal marker, with lower level clustering
# within individuals. The model includes only a random intercept
# at the individual-level, and then a random intercept and
# random slope at the lower cluster level. The association
# structure is based on a summation of the expected values for
# each of the lower level units.
simdat5 <- simjm(M = 1, n = 30,
                 random_trajectory = "none",
                 betaLong_intercept = -1,
                 b_sd = 1,
                 clust_control = list(
                   L = 6,
                   assoc = "sum",
                   random_trajectory = "linear",
                   u_sd = c(1,1)
                 ))

}
